无法报价功能添加

+++ pms/src/main/java/com/upedge/pms/modules/quote/dto/QuoteApplyListDto.java
@@ -34,7 +34,7 @@ public class QuoteApplyListDto {
     /**
      * 店铺产品ID
      */
-    private Integer storeProductId;
+    private Long storeProductId;
     /**
      * 店铺产品SKU
      */
